Broken publish step on CI
Summary:
Every `job` is run in a separate container and so, `checkout` step is required for the Git host to be added to `~/.ssh/known_hosts`. Without this step, it will timeout after 10 minutes waiting for you to add (yes) or reject (no) from known hosts (we get this prompt when we checkout from a host for the very first time).
Closes https://github.com/facebook/react-native/pull/17956

Differential Revision: D6968130

Pulled By: hramos

fbshipit-source-id: 6d62166fd375f8f408cf5f18b188f841d035d97f